refactor(ui,nav): extract NavGraph, simplify MainActivity; theme shapes; modernize bottom bar; polish auth, explore, recipe screens
