+++ app/src/main/java/com/example/cookit/ui/theme/Theme.kt
@@ -9,6 +9,7 @@ import androidx.compose.material3.dynamicLightColorScheme
 import androidx.compose.material3.lightColorScheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.material3.Shapes
 
 private val DarkColorScheme = darkColorScheme(
     primary = PrimaryColor,
@@ -64,6 +65,7 @@ fun CookITTheme(
     MaterialTheme(
         colorScheme = colorScheme,
         typography = Typography,
+        shapes = AppShapes,
         content = content
     )
 }
